окно проверки данных, скорректировал

+++ work_py/advanted_file.py
@@ -6,20 +6,20 @@ def skm_interval():
     str_raid = []
     if all([CreatePZ.dict_work_pervorations[plast]['отрайбировано'] == False for plast in CreatePZ.plast_work]):
         str_raid.append([CreatePZ.perforation_roof - 90, CreatePZ.perforation_roof - 10])
-        if CreatePZ.if_None(CreatePZ.paker_do['posle']) != 'отсут' and CreatePZ.if_None(CreatePZ.H_F_paker_do['posle']) != 'отсут' :
-            str_raid.append([CreatePZ.H_F_paker_do['posle'] - 20, CreatePZ.H_F_paker_do['posle'] + 20])
+        if CreatePZ.if_None(CreatePZ.paker_do["posle"]) != 'отсут' and CreatePZ.if_None(CreatePZ.H_F_paker_do["posle"]) != 'отсут' :
+            str_raid.append([CreatePZ.H_F_paker_do["posle"] - 20, CreatePZ.H_F_paker_do["posle"] + 20])
         if CreatePZ.leakiness:
 
             for nek in list(CreatePZ.dict_leakiness['НЭК']['интервал'].keys()):
                 print(f' наруш {nek}')
-                if int(nek.split('-')[1]) + 20 < CreatePZ.current_bottom:
-                    str_raid.append([int(nek.split('-')[0]) - 90, int(nek.split('-')[1]) + 20])
+                if float(nek.split('-')[1]) + 20 < CreatePZ.current_bottom:
+                    str_raid.append([int(float(nek.split('-')[0])) - 90, int(float(nek.split('-')[1])) + 20])
                 else:
-                    str_raid.append([int(nek.split('-')[0]) - 90,
+                    str_raid.append([int(float(nek.split('-')[0])) - 90,
                                      CreatePZ.CreatePZ.current_bottom - 2])
     else:
-        if CreatePZ.paker_do['posle'] != 'отсут':
-            str_raid.append([CreatePZ.H_F_paker_do['posle'] - 20, CreatePZ.H_F_paker_do['posle'] + 20])
+        if CreatePZ.paker_do["posle"] != 'отсут':
+            str_raid.append([float(CreatePZ.H_F_paker_do["posle"]) - 20, float(CreatePZ.H_F_paker_do["posle"]) + 20])
         if CreatePZ.leakiness:
             for nek in list(CreatePZ.dict_leakiness['НЭК']['интервал'].keys()):
                 if int(nek.split('-')[1]) + 20 < CreatePZ.current_bottom:
